// src/app/shared/interfaces/indicadores-infracoes.model.ts
export interface IndicadoresInfracoesMV {
  infracoesPeriodo?: {
    data: any[];
    filter: any[]
  };
  infracoesDiaSemana?: {
    data: any[];
    filter: any[]
  };
  infracoesHorario?: {
    data: any[];
    filter: any[]
  };
  infracoesHorarioComercial?: {
    data: any[];
    filter: any[]
  };
  principaisPlacas?: {
    data: any[];
    filter: any[]
  };
  principaisMotivos?: {
    data: any[];
    filter: any[]
  };
  infracoesCentroCusto?: {
    data: any[];
    filter: any[]
  };
  infracoesNIC?: {
    data: any[];
    filter: any[]
  };
  prazoIdentificacao?: {
    data: any[];
    filter: any[]
  };
  despesasInfracoes?: {
    data: any[];
    filter: any[]
  };
}